Revert "batch-system: add benchmark (#6477)"
This reverts commit 3971a3a2cec03fa6a122adea044582af46397599.
